agents/requirements.txt
fastapi
uvicorn[standard]
langchain
langchain-core
langchain-community
openai
python-dotenv
crawl4ai
playwright
requests
beautifulsoup4
nest_asyncio
